Address wobbly spin animation
Due to the parent element not being a square, the `.spin` class and subsequent animation was causing the refresh icon to not pivot correctly. Making the icon of the parent spin instead fixes this.
This change also means the refresh/"check thread" icon in the Thread Watcher now turns instead of remaining stationery. Closes #64